+++ client/packages/host/src/Help/Help.tsx
@@ -6,7 +6,7 @@ import {
 import React from 'react';
 import { AppVersion } from '../components';
 import { SiteInfo } from '../components/SiteInfo';
-import { FeedbackSection } from './FeedbackSection';
+import { ContactFormSection } from './ContactFormSection';
 import { UserGuide } from '../Admin/UserGuide';
 
 export const Help: React.FC = () => {
@@ -15,7 +15,7 @@ export const Help: React.FC = () => {
   return (
     <Box flex={1} padding={4} sx={{ maxWidth: 800 }}>
       <UserGuide />
-      <FeedbackSection />
+      <ContactFormSection />
       <AppBarButtonsPortal>
         <AppVersion SiteInfo={<SiteInfo siteName={initStatus?.siteName} />} />
       </AppBarButtonsPortal>
