--- client/packages/host/src/api/hooks/help/useFeedbackFormGraphQL.ts
@@ -1,16 +1,0 @@
-import { 
-    useAuthContext, 
-    useGql, 
-    useQueryClient, 
-} from '@openmsupply-client/common';
-import { getSdk } from '../../operations.generated';
-
-export const useFeedbackFormGraphQL = () => { 
-    const { client } = useGql();
-    const queryClient = useQueryClient();
-    const { storeId, store } = useAuthContext();
-    const api = getSdk(client);
-
-    return { api, queryClient, storeId, store };
-}
-
